Extract CoreData managed model classes into a separate files
